refactor(client): enhance error handling with Arc for retry status codes and introduce into_retry function
