+++ crates/forge_provider/src/lib.rs
@@ -2,6 +2,7 @@ mod anthropic;
 mod client;
 mod error;
 mod forge_provider;
+mod retry;
 mod utils;
 
 // Re-export from builder.rs
